requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==23.2.0
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
dnspython==2.7.0
elastic-transport==8.17.1
elasticsearch==8.17.2
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
jiter==0.9.0
kombu==5.5.2
multidict==6.4.3
openai==1.75.0
packaging==24.2
prompt-toolkit==3.0.50
propcache==0.3.1
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.9.1
pymongo==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyyaml==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tiktoken==0.9.0
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.19.0
